* Koans integration.
* Add KoansMarketWrapper

* Functionality is nearly the same as Nouns, but we change
  some lines in KoansMarketWrapper to account for the Koans'
  auction having 7 fields vs Nouns' 6.

* Tests are laregly the same.
