+++ hardhat.config.js
@@ -15,6 +15,15 @@ module.exports = {
   solidity: {
     compilers: [
       {
+        version: "0.8.6",
+        settings: {
+          optimizer: {
+            enabled: true,
+            runs: 10000,
+          },
+        },
+      },
+      {
         version: '0.8.5',
         settings: {
           optimizer: {
