+++ test/helpers/externalTransactions.js
@@ -9,6 +9,8 @@ async function placeBid(signer, marketContract, auctionId, value, marketName) {
         data = encodeData(marketContract, 'createBid', [auctionId]);
     } else if(marketName == MARKET_NAMES.FOUNDATION) {
         data = encodeData(marketContract, 'placeBid', [auctionId]);
+    } else if (marketName == MARKET_NAMES.KOANS) {
+        data = encodeData(marketContract, 'createBid', [auctionId]);
     } else {
         throw new Error("Unsupported Market");
     }
@@ -28,6 +30,8 @@ async function externalFinalize(signer, marketContract, auctionId, marketName) {
         data = encodeData(marketContract, 'settleCurrentAndCreateNewAuction', []);
     } else if(marketName == MARKET_NAMES.FOUNDATION) {
         data = encodeData(marketContract, 'finalizeReserveAuction', [auctionId]);
+    } else if (marketName == MARKET_NAMES.KOANS) {
+        data = encodeData(marketContract, 'settleCurrentAndCreateNewAuction', []);
     } else {
         throw new Error("Unsupported Market");
     }
